Bump to 2.1.0
This changes the minimum supported IDE version to 162 (2016.2).
This fixes a regression in which v2.0.0 bumped supported IDE to 2017.3.
This occurred because of updateSinceUntilBuild true in build.gradle.

Rather than reverting to 143+ as it was pre-2.0, I am sticking to the
minimum "IntelliJ Platform" product.

See http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html